% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bigQueryR.R
\docType{package}
\name{bigQueryR}
\alias{bigQueryR}
\alias{bigQueryR-package}
\title{bigQueryR}
\description{
Provides an interface with Google BigQuery
}
\seealso{
https://cloud.google.com/bigquery/docs/reference/v2/?hl=en
}

