% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{bq_get_global_project}
\alias{bq_get_global_project}
\title{Get global bucket name}
\usage{
bq_get_global_project()
}
\value{
Bucket name
}
\description{
Bucket name set this session to use by default
}
\details{
Set the bucket name via \link{gcs_global_bucket}
}
\seealso{
Other bucket functions: \code{\link{bq_global_project}}
}
