% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bqr_copy_table}
\alias{bqr_copy_table}
\title{Copy BigQuery table}
\usage{
bqr_copy_table(source_projectid, source_datasetid, source_tableid,
  destination_projectid, destination_datasetid, destination_tableid,
  createDisposition = c("CREATE_IF_NEEDED", "CREATE_NEVER"),
  writeDisposition = c("WRITE_TRUNCATE", "WRITE_APPEND", "WRITE_EMPTY"))
}
\arguments{
\item{source_projectid}{source table's projectId}

\item{source_datasetid}{source table's datasetId}

\item{source_tableid}{source table's tableId}

\item{destination_projectid}{destination table's projectId}

\item{destination_datasetid}{destination table's datasetId}

\item{destination_tableid}{destination table's tableId}

\item{createDisposition}{Create table's behaviour}

\item{writeDisposition}{Write to an existing table's behaviour}
}
\value{
A job object
}
\description{
Copy a source table to another destination
}
