% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bqr_create_table}
\alias{bqr_create_table}
\title{Create a Table}
\usage{
bqr_create_table(projectId, datasetId, tableId, template_data)
}
\arguments{
\item{projectId}{The BigQuery project ID.}

\item{datasetId}{A datasetId within projectId.}

\item{tableId}{Name of table you want.}

\item{template_data}{A dataframe with the correct types of data}
}
\value{
A table schema in R list form
}
\description{
Create a Table
}
\details{
A temporary csv file is creted if you choose a dataframe.
}

