% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bqr_delete_table}
\alias{bqr_delete_table}
\title{Delete a Table}
\usage{
bqr_delete_table(projectId = bq_get_global_project(),
  datasetId = bq_get_global_dataset(), tableId)
}
\arguments{
\item{projectId}{The BigQuery project ID.}

\item{datasetId}{A datasetId within projectId.}

\item{tableId}{Name of table you want to delete.}
}
\value{
TRUE if deleted, FALSE if not.
}
\description{
Delete a Table
}
\details{
Deletes a BigQuery table
}
\seealso{
Other bigQuery meta functions: \code{\link{bqr_create_table}},
  \code{\link{bqr_list_datasets}},
  \code{\link{bqr_list_projects}},
  \code{\link{bqr_list_tables}},
  \code{\link{bqr_table_data}},
  \code{\link{bqr_table_meta}}
}
