% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadData.R
\name{bqr_do_upload.character}
\alias{bqr_do_upload.character}
\title{upload for gs:// character vector}
\usage{
\method{bqr_do_upload}{character}(upload_data, projectId, datasetId, tableId,
  create, user_schema, sourceFormat)
}
\description{
upload for gs:// character vector
}
