% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadData.R
\name{bqr_do_upload.data.frame}
\alias{bqr_do_upload.data.frame}
\title{upload for local data.fram}
\usage{
\method{bqr_do_upload}{data.frame}(upload_data, projectId, datasetId, tableId,
  create, user_schema, sourceFormat)
}
\description{
upload for local data.fram
}
