% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadData.R
\name{bqr_do_upload}
\alias{bqr_do_upload}
\title{S3 generic dispatch}
\usage{
bqr_do_upload(upload_data, projectId, datasetId, tableId, create, user_schema,
  sourceFormat)
}
\description{
S3 generic dispatch
}
