% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{bqr_download_extract}
\alias{bqr_download_extract}
\title{Download extract data}
\usage{
bqr_download_extract(extractJob, filename = NULL)
}
\arguments{
\item{extractJob}{An extract job from \link{bqr_extract_data}}

\item{filename}{Where to save the csv file. If NULL then uses objectname.}
}
\value{
TRUE if successfully downloaded
}
\description{
After extracting data via \link{bqr_extract_data} download the 
  extract from the Google Storage bucket.
  
If more than 1GB, will save multiple .csv files with prefix "N_" to filename.
}
\seealso{
Other BigQuery asynch query functions: 
\code{\link{bqr_extract_data}()},
\code{\link{bqr_get_job}()},
\code{\link{bqr_grant_extract_access}()},
\code{\link{bqr_query_asynch}()},
\code{\link{bqr_wait_for_job}()}
}
\concept{BigQuery asynch query functions}
