% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{bqr_get_global_dataset}
\alias{bqr_get_global_dataset}
\alias{bq_get_global_dataset}
\title{Get global dataset name}
\usage{
bqr_get_global_dataset()

bq_get_global_dataset()
}
\value{
dataset name
}
\description{
dataset name set this session to use by default
}
\details{
Set the dataset name via \link{bq_global_dataset}
}
\seealso{
Other dataset functions: \code{\link{bqr_global_dataset}}
}
\concept{dataset functions}
