% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{bqr_get_job}
\alias{bqr_get_job}
\title{Poll a jobId}
\usage{
bqr_get_job(projectId, jobId)
}
\arguments{
\item{projectId}{projectId of job}

\item{jobId}{jobId to poll}
}
\value{
A Jobs resource
}
\description{
Poll a jobId
}

