% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{bqr_global_dataset}
\alias{bqr_global_dataset}
\alias{bq_global_dataset}
\title{Set global dataset name}
\usage{
bqr_global_dataset(dataset)

bq_global_dataset(dataset)
}
\arguments{
\item{dataset}{dataset name you want this session to use by default, or a dataset object}
}
\value{
The dataset name (invisibly)
}
\description{
Set a dataset name used for this R session
}
\details{
This sets a dataset to a global environment value so you don't need to
supply the dataset argument to other API calls.
}
\seealso{
Other dataset functions: 
\code{\link{bqr_get_global_dataset}()}
}
\concept{dataset functions}
