% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{bqr_global_project}
\alias{bqr_global_project}
\alias{bq_global_project}
\title{Set global project name}
\usage{
bqr_global_project(project)

bq_global_project(...)
}
\arguments{
\item{project}{project name you want this session to use by default, or a project object}
}
\value{
The project name (invisibly)
}
\description{
Set a project name used for this R session
}
\details{
This sets a project to a global environment value so you don't need to
supply the project argument to other API calls.
}
\seealso{
Other project functions: \code{\link{bqr_get_global_project}}
}
