% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{bqr_grant_extract_access}
\alias{bqr_grant_extract_access}
\title{Grant access to an extract on Google Cloud Storage}
\usage{
bqr_grant_extract_access(extractJob, email)
}
\arguments{
\item{extractJob}{An extract job from \link{bqr_extract_data}}

\item{email}{email of the user to have access}
}
\value{
URL(s) to download the extract that is accessible by email
}
\description{
To access the data created in \link{bqr_extract_data}.
Requires the Google account email of the user.
}
\details{
Uses \href{https://cloud.google.com/storage/docs/authentication#cookieauth}{cookie based auth}.
}

