% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{bqr_list_jobs}
\alias{bqr_list_jobs}
\title{List BigQuery jobs}
\usage{
bqr_list_jobs(projectId = bqr_get_global_project(), allUsers = FALSE,
  projection = c("full", "minimal"), stateFilter = c("done", "pending",
  "running"))
}
\arguments{
\item{projectId}{projectId of job}

\item{allUsers}{Whether to display jobs owned by all users in the project.}

\item{projection}{"full" - all job data, "minimal" excludes job configuration.}

\item{stateFilter}{Filter for job status.}
}
\value{
A list of jobs resources
}
\description{
List the BigQuery jobs for the projectId
}
\details{
Lists all jobs that you started in the specified project. 
Job information is available for a six month period after creation. 
The job list is sorted in reverse chronological order, by job creation time. 
Requires the Can View project role, or the 
  Is Owner project role if you set the allUsers property.
}
