% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBigQuery.R
\name{bqr_list_projects}
\alias{bqr_list_projects}
\title{List Google Dev Console projects you have access to}
\usage{
bqr_list_projects()
}
\value{
A dataframe of the projects you have access to under the authentication
}
\description{
Example: bqr_list_projects()
}
\examples{

\dontrun{
  library(bigQueryR)
  
  ## this will open your browser
  ## Authenticate with an email that has access to the BigQuery project you need
  bqr_auth()
  
  ## verify under a new user
  bqr_auth(new_user=TRUE)
  
  ## get projects
  projects <- bqr_list_projects()
  
}

}
\seealso{
Other bigQuery meta functions: \code{\link{bqr_create_table}},
  \code{\link{bqr_delete_table}},
  \code{\link{bqr_list_datasets}},
  \code{\link{bqr_list_tables}},
  \code{\link{bqr_patch_table}},
  \code{\link{bqr_table_data}},
  \code{\link{bqr_table_meta}}
}
\concept{bigQuery meta functions}
