% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{bqr_query}
\alias{bqr_query}
\title{Query a BigQuery Table}
\usage{
bqr_query(projectId, datasetId, query, maxResults = 1000)
}
\arguments{
\item{projectId}{The BigQuery project ID}

\item{datasetId}{A datasetId within projectId}

\item{query}{BigQuery SQL}

\item{maxResults}{Max number per page of results. Set total rows with LIMIT in your query.}
}
\value{
a data.frame. 
  If there is an SQL error, a data.frame with 
  additional class "bigQueryR_query_error" and the 
  problem in the data.frame$message
}
\description{
MaxResults is how many results to return per page of results, which can be less than the 
total results you have set in your  query using LIMIT.  Google recommends for bigger datasets
to set maxResults = 1000, but this will use more API calls.

Example: 
bqr_query("big-query-r","samples",
          "SELECT COUNT(repository.url) FROM [publicdata:samples.github_nested]")
          
SELECT COUNT(repository.url) as freq, repository.language FROM
[publicdata:samples.github_nested]
GROUP BY repository.language
ORDER BY freq DESC
LIMIT
1000
}

