% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{bqr_query_asynch}
\alias{bqr_query_asynch}
\title{BigQuery query asynchronously}
\usage{
bqr_query_asynch(projectId, datasetId, query, destinationTableId,
  writeDisposition = c("WRITE_EMPTY", "WRITE_TRUNCATE", "WRITE_APPEND"))
}
\arguments{
\item{projectId}{projectId to be billed.}

\item{datasetId}{datasetId of where query will execute.}

\item{query}{The BigQuery query as a string.}

\item{destinationTableId}{Id of table the results will be written to.}

\item{writeDisposition}{Behaviour if destination table exists. See Details.}
}
\value{
A Job object to be queried via \link{bqr_get_job}
}
\description{
Use for big results > 10000 that write to their own destinationTableId
}
\details{
writeDisposition - behaviour if destinationTable already exists: 
    WRITE_TRUNCATE: BigQuery overwrites the table data. 
    WRITE_APPEND: BigQuery appends the data to the table
    WRITE_EMPTY: If contains data, a 'duplicate' error is returned
}

