% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{bqr_signed_url}
\alias{bqr_signed_url}
\title{Create a signed URL for Google Cloud Storage}
\usage{
bqr_signed_url(extractJob, expiration = 86400, pem, email)
}
\arguments{
\item{extractJob}{An extract job from \link{bqr_extract_data}}

\item{expiration}{How long the link is valid for in seconds}

\item{pem}{file path to a .pem service account key}

\item{email}{email of the Google service account}
}
\value{
Signed URL(s)
}
\description{
To access the data created in \link{bqr_extract_data}.
Requires a service account .pem:
}
\details{
1. Create a service account and download .p12 key
  2. Run in console: 
  openssl pkcs12 -in path/to/key.p12 -nodes -nocerts > path/to/key.pem
}

