% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{bqr_table_data}
\alias{bqr_table_data}
\title{Get BigQuery Table's data list}
\usage{
bqr_table_data(projectId, datasetId, tableId, maxResults = 1000)
}
\arguments{
\item{projectId}{The BigQuery project ID}

\item{datasetId}{A datasetId within projectId}

\item{tableId}{The tableId within the datasetId

Not very useful as can't deal with nested datasets}
}
\description{
Get BigQuery Table's data list
}

