% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{bqr_table_meta}
\alias{bqr_table_meta}
\title{Get BigQuery Table meta data}
\usage{
bqr_table_meta(projectId, datasetId, tableId)
}
\arguments{
\item{projectId}{The BigQuery project ID}

\item{datasetId}{A datasetId within projectId}

\item{tableId}{The tableId within the datasetId

Example: bqr_table_meta("publicdata", "samples", "github_nested")}
}
\description{
Get BigQuery Table meta data
}

