% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{bqr_wait_for_job}
\alias{bqr_wait_for_job}
\title{Wait for a bigQuery job}
\usage{
bqr_wait_for_job(job, wait = 5)
}
\arguments{
\item{wait}{The number of seconds to wait between checks

Use this function to do a loop to check progress of a job running}

\item{jobObject}{A job object}
}
\value{
After a while, a completed job
}
\description{
Wait for a bigQuery job to finish.
}
\seealso{
Other BigQuery asynch query functions: \code{\link{bqr_download_extract}},
  \code{\link{bqr_extract_data}},
  \code{\link{bqr_get_job}},
  \code{\link{bqr_grant_extract_access}},
  \code{\link{bqr_query_asynch}}
}

