% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadData.R
\name{schema_fields}
\alias{schema_fields}
\title{Create data schema for upload to BigQuery}
\usage{
schema_fields(data)
}
\arguments{
\item{data}{An example of the data to create a schema from}
}
\value{
A schema object suitable to pass within the \code{schema} argument of \link{bqr_upload_data}
}
\description{
Use this on a sample of the data you want to load from Google Cloud Storage
}
\details{
This is taken from \link[bigrquery]{insert_upload_job}
}
\author{
Hadley Wickham \email{hadley@rstudio.com}
}
