% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{bq_get_global_project}
\alias{bq_get_global_project}
\title{Get global project name}
\usage{
bq_get_global_project()
}
\value{
project name
}
\description{
project name set this session to use by default
}
\details{
Set the project name via \link{gcs_global_project}
}
\seealso{
Other project functions: \code{\link{bq_global_project}}
}
