% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bqr_copy_table}
\alias{bqr_copy_table}
\title{Copy BigQuery table}
\usage{
bqr_copy_table(source_tableid, destination_tableid,
  source_projectid = bqr_get_global_project(),
  source_datasetid = bqr_get_global_dataset(),
  destination_projectid = bqr_get_global_project(),
  destination_datasetid = bqr_get_global_dataset(),
  createDisposition = c("CREATE_IF_NEEDED", "CREATE_NEVER"),
  writeDisposition = c("WRITE_TRUNCATE", "WRITE_APPEND", "WRITE_EMPTY"))
}
\arguments{
\item{source_tableid}{source table's tableId}

\item{destination_tableid}{destination table's tableId}

\item{source_projectid}{source table's projectId}

\item{source_datasetid}{source table's datasetId}

\item{destination_projectid}{destination table's projectId}

\item{destination_datasetid}{destination table's datasetId}

\item{createDisposition}{Create table's behaviour}

\item{writeDisposition}{Write to an existing table's behaviour}
}
\value{
A job object
}
\description{
Copy a source table to another destination
}
\seealso{
Other Table meta functions: \code{\link{Table}},
  \code{\link{bqr_create_table}},
  \code{\link{bqr_delete_table}},
  \code{\link{bqr_list_tables}},
  \code{\link{bqr_patch_table}},
  \code{\link{bqr_table_data}},
  \code{\link{bqr_table_meta}}
}
\concept{Table meta functions}
