% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bqr_create_table}
\alias{bqr_create_table}
\title{Create a Table}
\usage{
bqr_create_table(projectId, datasetId, tableId, template_data)
}
\arguments{
\item{projectId}{The BigQuery project ID.}

\item{datasetId}{A datasetId within projectId.}

\item{tableId}{Name of table you want.}

\item{template_data}{A dataframe with the correct types of data}
}
\value{
TRUE if created, FALSE if not.
}
\description{
Create a Table
}
\details{
Creates a BigQuery table
}
\seealso{
Other bigQuery meta functions: \code{\link{bqr_auth}},
  \code{\link{bqr_delete_table}},
  \code{\link{bqr_list_datasets}},
  \code{\link{bqr_list_projects}},
  \code{\link{bqr_list_tables}},
  \code{\link{bqr_table_data}},
  \code{\link{bqr_table_meta}}
}

