% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{bqr_list_datasets}
\alias{bqr_list_datasets}
\title{List BigQuery datasets}
\usage{
bqr_list_datasets(projectId)
}
\arguments{
\item{projectId}{The BigQuery project ID}
}
\description{
List BigQuery datasets
}

