% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bqr_list_tables}
\alias{bqr_list_tables}
\title{List BigQuery tables in a dataset}
\usage{
bqr_list_tables(projectId, datasetId)
}
\arguments{
\item{projectId}{The BigQuery project ID}

\item{datasetId}{A datasetId within projectId

Example: bqr_list_tables("publicdata", "samples")}
}
\description{
List BigQuery tables in a dataset
}

