% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bqr_patch_table}
\alias{bqr_patch_table}
\title{Update a Table}
\usage{
bqr_patch_table(Table)
}
\arguments{
\item{Table}{A Table object as created by \link{Table}}
}
\description{
This uses PATCH semantics to alter an existing table.  
 You need to create the Table object first to pass in using \link{Table} 
which will be transformed to JSON
}
\seealso{
\href{https://cloud.google.com/bigquery/docs/reference/rest/v2/tables#resource}{Definition of tables}

Other Table meta functions: 
\code{\link{Table}()},
\code{\link{bqr_copy_table}()},
\code{\link{bqr_create_table}()},
\code{\link{bqr_delete_table}()},
\code{\link{bqr_list_tables}()},
\code{\link{bqr_table_data}()},
\code{\link{bqr_table_meta}()}
}
\concept{Table meta functions}
