% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{bqr_query}
\alias{bqr_query}
\title{Query a BigQuery Table}
\usage{
bqr_query(projectId, datasetId, tableId, query, maxResults = 1000)
}
\arguments{
\item{projectId}{The BigQuery project ID}

\item{datasetId}{A datasetId within projectId}

\item{tableId}{The tableId within the datasetId}

\item{query}{BigQuery SQL}

\item{MaxResults}{Max number of results

Example:
bqr_query("big-query-r","samples","github_nested",
          "SELECT COUNT(repository.url) FROM [publicdata:samples.github_nested]")}
}
\description{
Query a BigQuery Table
}

