% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bqr_table_data}
\alias{bqr_table_data}
\title{Get BigQuery Table's data list}
\usage{
bqr_table_data(projectId = bqr_get_global_project(),
  datasetId = bqr_get_global_dataset(), tableId, maxResults = 1000)
}
\arguments{
\item{projectId}{The BigQuery project ID}

\item{datasetId}{A datasetId within projectId}

\item{tableId}{The tableId within the datasetId}

\item{maxResults}{Number of results to return}
}
\value{
data.frame of table data

This won't work with nested datasets, for that use \link{bqr_query} as that flattens results.
}
\description{
Get BigQuery Table's data list
}
\seealso{
Other Table meta functions: \code{\link{Table}},
  \code{\link{bqr_copy_table}},
  \code{\link{bqr_create_table}},
  \code{\link{bqr_delete_table}},
  \code{\link{bqr_list_tables}},
  \code{\link{bqr_patch_table}},
  \code{\link{bqr_table_meta}}
}
\concept{Table meta functions}
