% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bqr_table_meta}
\alias{bqr_table_meta}
\title{Get BigQuery Table meta data}
\usage{
bqr_table_meta(projectId = bq_get_global_project(),
  datasetId = bq_get_global_dataset(), tableId)
}
\arguments{
\item{projectId}{The BigQuery project ID}

\item{datasetId}{A datasetId within projectId}

\item{tableId}{The tableId within the datasetId}
}
\value{
list of table metadata
}
\description{
Get BigQuery Table meta data
}
\examples{

\dontrun{
  bqr_table_meta("publicdata", "samples", "github_nested")
}


}
\seealso{
Other bigQuery meta functions: \code{\link{bqr_create_table}},
  \code{\link{bqr_delete_table}},
  \code{\link{bqr_list_datasets}},
  \code{\link{bqr_list_projects}},
  \code{\link{bqr_list_tables}},
  \code{\link{bqr_table_data}}
}
